requirements.txt
Flask==1.1.2
PyMySQL==0.9.3
Flask-SQLAlchemy==2.4.3
uWSGI==2.0.19.1
Flask-Login==0.5.0
SQLAlchemy==1.3.17
argon2-cffi==20.1.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Session==0.3.2
isort==4.3.21